"Added weather forecast script to get weather information about a city"
This commit adds a Python script to get weather infro about a city using the OpenWeatherMap API. This program is strictly designed to be used as a module. However, you can still modify it according to your need.
Make sure to go through the code if you want to use it in your program.