"Added a Python script to get quotes from zenquotes API"
This commit adds a Python script to get quotes from zenquotes API. It can generate a random quote, quote of the day or n number of quotes (n =< 50). No API key is required to run the program. But make sure to go through the code if you want to use it in your program.